+++ .dialyzer_ignore.exs
@@ -7,5 +7,6 @@
   ~r/lib\/flop\/schema\.ex.*no_return Function filterable\/1 has no local return\./,
   ~r/lib\/flop\/schema\.ex.*no_return Function max_limit\/1 has no local return\./,
   ~r/lib\/flop\/schema\.ex.*no_return Function pagination_types\/1 has no local return\./,
-  ~r/lib\/flop\/schema\.ex.*no_return Function sortable\/1 has no local return\./
+  ~r/lib\/flop\/schema\.ex.*no_return Function sortable\/1 has no local return\./,
+  {"lib/flop/filter.ex:193:unused_fun Function field_type_from_flop_schema/3 will never be called."}
 ]