+++ test/flop_test.exs
@@ -187,25 +187,11 @@ defmodule FlopTest do
       end
     end
 
-    @tag :capture_log
-    property "applies equality filter to compound fields" do
-      check all pet_count <- integer(@pet_count_range),
-                pets = insert_list_and_sort(pet_count, :pet_with_owner),
-                # only compound fields
-                field <- member_of([:full_name, :pet_and_owner_name]),
-                pet <- member_of(pets),
-                query_value <-
-                  pet
-                  |> Pet.concatenated_value_for_compound_field(field)
-                  |> constant(),
-                query_value != "" do
-        expected = filter_pets(pets, field, :==, query_value)
-
-        assert query_pets_with_owners(%{
-                 filters: [%{field: field, op: :==, value: query_value}]
-               }) == expected
-
-        checkin_checkout()
+    test "does not allow equality filter on compound fields" do
+      assert_raise Flop.InvalidParamsError, fn ->
+        query_pets_with_owners(%{
+          filters: [%{field: :full_name, op: :==, value: "o"}]
+        })
       end
     end
 
@@ -228,23 +214,11 @@ defmodule FlopTest do
       end
     end
 
-    @tag :capture_log
-    property "applies inequality filter to compound fields" do
-      check all pet_count <- integer(@pet_count_range),
-                pets = insert_list_and_sort(pet_count, :pet_with_owner),
-                # only compound fields
-                field <- member_of([:full_name, :pet_and_owner_name]),
-                pet <- member_of(pets),
-                query_value =
-                  Pet.concatenated_value_for_compound_field(pet, field),
-                query_value != "" do
-        expected = filter_pets(pets, field, :!=, query_value)
-
-        assert query_pets_with_owners(%{
-                 filters: [%{field: field, op: :!=, value: query_value}]
-               }) == expected
-
-        checkin_checkout()
+    test "does not allow inequality filter on compound fields" do
+      assert_raise Flop.InvalidParamsError, fn ->
+        query_pets_with_owners(%{
+          filters: [%{field: :full_name, op: :!=, value: "o"}]
+        })
       end
     end