+++ CHANGELOG.md
@@ -2,12 +2,23 @@
 
 ## Unreleased
 
-## Added
+### Added
 
 - Support for custom fields. These fields allow you to run custom filter
   functions for anything that cannot be expressed with Flop filters.
 - Added `fetch`, `get`, `get_all`, `delete`, `delete_first`, `drop`, `new`,
   `take`, `pop`, `pop_first`, `put` and `put_new` functions to `Flop.Filter`.
+- Added `ecto_type` option to join fields.
+
+### Changed
+
+- `Flop.Filter.allowed_operators/2` now tries to determine the Ecto type by
+  reading the Flop field type from the schema module. This function is used
+  during parameter validation, which means the validation step will be a bit
+  stricter now. For join and custom fields, the Ecto type is determined via the
+  new `ecto_type` option. If the option is not set, the function returns all
+  operators as before. For compound fields, only the supported operators are
+  returned.
 
 ## [0.18.4] - 2022-11-17